made dijkstra more flexible

+++ cities/destination.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass
 class Destination:
     destination: str
     length: float
-    morning_time: int
-    noon_time: int
-    evening_time: int
+    morning_time: float
+    noon_time: float
+    evening_time: float
 
     def night_time(self) -> float:
         return 75/100 * self.noon_time
